< :truck: & :building_construction: >[(fix/tree-structure)/arboresence] Update the files and folders tree structure for better lisibility and long terme developement
